requirements.txt
mypy==0.812
mypy-extensions==0.4.3
python-dotenv==0.16.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
